fix pulse sweep timer

+++ gameboy/include/gameboy/apu/data/sweep.h
@@ -19,7 +19,7 @@ struct sweep {
 
     bool enabled = false;
 
-    uint8_t timer = 0;
+    int16_t timer = 0;
     uint16_t shadow = 0u;
 
     [[nodiscard]] uint8_t sweep_count() const noexcept { return (reg.value() >> 4u) & 0x7u; }
